In scripts/make_bu_data, add some judgement

+++ scripts/make_bu_data.py
@@ -21,24 +21,22 @@ parser.add_argument('--output_dir', default='data/cocobu', help='output feature 
 args = parser.parse_args()
 csv.field_size_limit(sys.maxsize)
 
-
 FIELDNAMES = ['image_id', 'image_w','image_h','num_boxes', 'boxes', 'features']
-infiles = ['karpathy_test_resnet101_faster_rcnn_genome.tsv',
-          'karpathy_val_resnet101_faster_rcnn_genome.tsv',\
-          'karpathy_train_resnet101_faster_rcnn_genome.tsv.0', \
-           'karpathy_train_resnet101_faster_rcnn_genome.tsv.1']
-
 
 #----for my local set----#
-args.downloaded_feats = '/media/samsumg_1tb/Image_Caption/Datasets/MSCOCO/detection_features/trainval_2-10'
-args.output_dir = '/media/samsumg_1tb/Image_Caption/Datasets/MSCOCO/detection_features/trainval_2-10/trainval'
-# infiles =['trainval_resnet101_faster_rcnn_genome_36.tsv']
+infiles = ['test2014_resnet101_faster_rcnn_genome.tsv.0',
+          'test2014_resnet101_faster_rcnn_genome.tsv.1',\
+          'test2014_resnet101_faster_rcnn_genome.tsv.2']
+args.downloaded_feats = '/home/wzn/Datasets/ImageCaption/MSCOCO/detection_features/test2014_20-100'
+args.output_dir = '/home/wzn/Datasets/ImageCaption/MSCOCO/detection_features/test2014_20-100/test'
 #----for my local set----#
 
-
-os.makedirs(args.output_dir+'_att')
-os.makedirs(args.output_dir+'_fc')
-os.makedirs(args.output_dir+'_box')
+if not os.path.exists(args.output_dir+'_att'):
+    os.makedirs(args.output_dir+'_att')
+if not os.path.exists(args.output_dir+'_fc'):
+    os.makedirs(args.output_dir+'_fc')
+if not os.path.exists(args.output_dir+'_box'):
+    os.makedirs(args.output_dir+'_box')
 
 for infile in infiles:
     print('Reading ' + infile)
@@ -46,14 +44,12 @@ for infile in infiles:
         reader = csv.DictReader(tsv_in_file, delimiter='\t', fieldnames = FIELDNAMES)
         for item in reader:
             item['image_id'] = int(item['image_id'])
+            if os.path.exists(os.path.join(args.output_dir + '_att', str(item['image_id']))+'.npz'):
+                continue
             item['num_boxes'] = int(item['num_boxes'])
             for field in ['boxes', 'features']:
-                item[field] = np.frombuffer(base64.decodestring(item[field]), 
+                item[field] = np.frombuffer(base64.decodestring(item[field]),
                         dtype=np.float32).reshape((item['num_boxes'],-1))
             np.savez_compressed(os.path.join(args.output_dir+'_att', str(item['image_id'])), feat=item['features'])
             np.save(os.path.join(args.output_dir+'_fc', str(item['image_id'])), item['features'].mean(0))
             np.save(os.path.join(args.output_dir+'_box', str(item['image_id'])), item['boxes'])
-
-
-
-
